fixed to pass all cypress tests
