+++ src/components/Article.jsx
@@ -1,19 +1,40 @@
 import React from 'react';
-// eslint-disable-next-line import/no-cycle
-import { formatDate as format } from '../App';
 
-export const Article = ({ title, date, text }) => (
-  <div className="box">
-    <h2 className="title" data-cy="title">
-      {title}
-    </h2>
-    <p className="subtitle">
-      <time dateTime={format(date).dateISO} data-cy="date">
-        {format(date).dateLocaleString}
-      </time>
-    </p>
-    <p data-cy="text">
-      {text}
-    </p>
-  </div>
-);
+const formatDate = (date) => {
+  const today = new Date(date);
+  const dateISO = today.toISOString().slice(0, 10);
+  const dateLocaleString = today.toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  });
+
+  return {
+    dateISO,
+    dateLocaleString,
+  };
+};
+
+export const Article = (props) => {
+  const {
+    title,
+    date,
+    text,
+  } = props;
+
+  return (
+    <div className="box">
+      <h2 className="title" data-cy="title">
+        {title}
+      </h2>
+      <p className="subtitle">
+        <time dateTime={formatDate(date).dateISO} data-cy="date">
+          {formatDate(date).dateLocaleString}
+        </time>
+      </p>
+      <p data-cy="text">
+        {text}
+      </p>
+    </div>
+  );
+};
